refactor: route.js 유틸 폴더로  이동
